requirements.txt
APScheduler==3.6.3
asgiref==3.3.1
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.12.5
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
Django==3.0.5
django-cors-headers==3.6.0
django-rest-swagger==2.2.0
djangorestframework==3.12.2
djongo==1.3.3
dnspython==2.1.0
gunicorn==20.0.4
idna==2.10
itypes==1.2.0
Jinja2==2.11.2
jsonfield==3.1.0
MarkupSafe==1.1.1
openapi-codec==1.3.2
pymongo==3.11.2
python-decouple==3.4
pytz==2020.5
requests==2.25.1
simplejson==3.17.2
six==1.15.0
soupsieve==2.1
sqlparse==0.2.4
tzlocal==2.1
uritemplate==3.0.1
urllib3==1.26.2
waitress==1.4.4
